+++ packages/slider/src/Slider.tsx
@@ -7,7 +7,7 @@ export type GenericObject = { [key: string]: any };
 
 export interface SliderProps extends React.HTMLAttributes<HTMLDivElement> {
   onSubmit(values: GenericObject | Promise<GenericObject>): Promise<GenericObject> | GenericObject;
-  afterSubmit?(values: GenericObject | Promise<GenericObject>): Promise<GenericObject> | GenericObject;
+  afterSubmit?(values: GenericObject | Promise<GenericObject>): any;
   beforeSubmit?(values: GenericObject | Promise<GenericObject>): Promise<GenericObject> | GenericObject;
   afterSlideChange?(): void;
   autoComplete?: boolean;
@@ -95,6 +95,20 @@ export class Slider extends React.PureComponent<SliderProps, SliderState> {
 
   componentDidMount() {
     this.mounted = true;
+
+    /**
+     * Call any didEnter hooks on the first slide
+     */
+    const { didEnter, didEnterAsNext } = this.currentSlide.props;
+
+    if (didEnterAsNext) {
+      execIfFunc(didEnterAsNext, this.injectSlideProps());
+      return;
+    }
+
+    if (didEnter) {
+      execIfFunc(didEnter, this.injectSlideProps());
+    }
   }
 
   componentDidUpdate(_: SliderProps, prevState: SliderState) {